API

Public

Kroki.Kroki — Module

The main Module containing the necessary types of functions for integration with a Kroki service.

Defines Base.show and corresponding Base.showable methods for different output formats and Diagram types, so they render in their most optimal form in different environments (e.g. the documentation system, Documenter output, Jupyter, etc.).

Kroki.Diagram — Type

struct Diagram

A representation of a diagram that can be rendered by a Kroki service.

Examples

julia> Kroki.Diagram(:PlantUML, "Kroki -> Julia: Hello Julia!")
     ┌─────┐          ┌─────┐
     │Kroki│          │Julia│
     └──┬──┘          └──┬──┘
        │ Hello Julia!   │
        │───────────────>│
     ┌──┴──┐          ┌──┴──┐
     │Kroki│          │Julia│
     └─────┘          └─────┘

Fields

specification::AbstractString

The textual specification of the diagram

type::Symbol

The type of diagram specification (e.g. ditaa, Mermaid, PlantUML, etc.). This value is case-insensitive.

Service Management

Kroki.Service — Module

Defines functions and constants managing the Kroki service the rest of the package uses to render diagrams. These services can be either local or remote.

This module also enables management of a local service instance, provided Docker and Docker Compose are available on the system.

Exports

Imports

Base

Core

DocStringExtensions

Kroki.Documentation

Kroki.Service.DockerComposeExecutionError — Type

struct DockerComposeExecutionError <: Exception

A specialized Exception to include reporting instructions for specific types of errors that may occur while trying to execute docker-compose.

Fields

message::String

Kroki.Service.setEndpoint! — Function

setEndpoint!() -> String
setEndpoint!(endpoint::AbstractString) -> String

Sets the ENDPOINT using a fallback mechanism if no endpoint is provided.

The fallback mechanism checks for a KROKI_ENDPOINT environment variable specifying an endpoint (e.g. to be used across Julia instances). If this environment variable is also not present the DEFAULT_ENDPOINT is used.

This can, for instance, be used in cases where a privately hosted instance is available or when a local service has been start!ed.

Returns the value that ENDPOINT got set to.

Examples

setEndpoint!()

setEndpoint!("http://localhost:8000")

Kroki.Service.start! — Function

start!()
start!(update_endpoint::Bool)

Starts the Kroki service components on the local system, optionally, ensuring ENDPOINT points to them.

Pass false to the function to prevent the ENDPOINT from being updated. The default behavior is to update.

Kroki.Service.status — Method

status() -> Any

Returns a NamedTuple where the keys are the names of the service components and the values their corresponding 'running' state.

Examples

julia> status()
(core = true, mermaid = false)

Kroki.Service.stop! — Function

stop!()
stop!(perform_cleanup::Bool)

Stops any running Kroki service components ensuring ENDPOINT no longer points to the stopped service.

Cleans up left-over containers by default. This behavior can be turned off by passing false to the function.

Kroki.Service.update! — Method

update!()

Updates the Docker images for the individual Kroki service components.

Shorthands

Kroki.@actdiag_str — Macro

Shorthand for instantiating actdiag Diagrams.

Kroki.@blockdiag_str — Macro

Shorthand for instantiating blockdiag Diagrams.

Kroki.@bpmn_str — Macro

Shorthand for instantiating bpmn Diagrams.

Kroki.@bytefield_str — Macro

Shorthand for instantiating bytefield Diagrams.

Kroki.@c4plantuml_str — Macro

Shorthand for instantiating c4plantuml Diagrams.

Kroki.@ditaa_str — Macro

Shorthand for instantiating ditaa Diagrams.

Kroki.@erd_str — Macro

Shorthand for instantiating erd Diagrams.

Kroki.@excalidraw_str — Macro

Shorthand for instantiating excalidraw Diagrams.

Kroki.@graphviz_str — Macro

Shorthand for instantiating graphviz Diagrams.

Kroki.@mermaid_str — Macro

Shorthand for instantiating mermaid Diagrams.

Kroki.@nomnoml_str — Macro

Shorthand for instantiating nomnoml Diagrams.

Kroki.@nwdiag_str — Macro

Shorthand for instantiating nwdiag Diagrams.

Kroki.@packetdiag_str — Macro

Shorthand for instantiating packetdiag Diagrams.

Kroki.@pikchr_str — Macro

Shorthand for instantiating pikchr Diagrams.

Kroki.@plantuml_str — Macro

Shorthand for instantiating plantuml Diagrams.

Kroki.@rackdiag_str — Macro

Shorthand for instantiating rackdiag Diagrams.

Kroki.@seqdiag_str — Macro

Shorthand for instantiating seqdiag Diagrams.

Kroki.@structurizr_str — Macro

Shorthand for instantiating structurizr Diagrams.

Kroki.@svgbob_str — Macro

Shorthand for instantiating svgbob Diagrams.

Kroki.@umlet_str — Macro

Shorthand for instantiating umlet Diagrams.

Kroki.@vega_str — Macro

Shorthand for instantiating vega Diagrams.

Kroki.@vegalite_str — Macro

Shorthand for instantiating vegalite Diagrams.

Kroki.@wavedrom_str — Macro

Shorthand for instantiating wavedrom Diagrams.

Private

Kroki.InvalidDiagramSpecificationError — Type

struct InvalidDiagramSpecificationError <: Exception

An Exception to be thrown when a Diagram representing an invalid specification is passed to render.

Fields

error::String

cause::Kroki.Diagram

Kroki.InvalidOutputFormatError — Type

struct InvalidOutputFormatError <: Exception

An Exception to be thrown when a Diagram is rendered to an unsupported or invalid output format.

Fields

error::String

cause::Kroki.Diagram

Kroki.LIMITED_DIAGRAM_SUPPORT — Constant

Some MIME types are not supported by all diagram types, this constant contains all these limitations. The union of all values corresponds to all supported Diagram types.

Kroki.UriSafeBase64Payload — Function

UriSafeBase64Payload(diagram::Kroki.Diagram) -> String

Compresses a Diagram's specification using zlib, turning the resulting bytes into a URL-safe Base64 encoded payload (i.e. replacing + by - and / by _) to be used in communication with a Kroki service.

See the Kroki documentation for more information.

Kroki.render — Function

render(diagram::Kroki.Diagram, output_format::AbstractString) -> Array{UInt8,1}

Renders a Diagram through a Kroki service to the specified output format.

If the Kroki service responds with an error throws an InvalidDiagramSpecificationError or InvalidOutputFormatError if a know type of error occurs. Other errors (e.g. HTTP.ExceptionRequest.StatusError for connection errors) are propagated if they occur.

Documentation

Kroki.Documentation — Module

Contains templates and a helper macro @setupDocstringMarkup to easily set up consistent docstring formats across modules.

Kroki.Documentation.@setupDocstringMarkup — Macro

Helper macro ensuring consistent docstring markup across modules through templating.

Service Management

Kroki.Service.DEFAULT_ENDPOINT — Constant

The default ENDPOINT to use, i.e. the publicly hosted version.

Kroki.Service.ENDPOINT — Constant

The currently active Kroki service endpoint being used.

Kroki.Service.SERVICE_DEFINITION_FILE — Constant

Path to the Docker Compose definitions for running a local Kroki service.

Kroki.Service.executeDockerCompose — Function

executeDockerCompose(cmd::Array{String,1}) -> String

Helper function for executing Docker Compose commands.

Returns captured stdout.

Throws an ErrorException if Docker and/or Docker Compose aren't available. Throws a DockerComposeExecutionError if any other exception occurs during execution.