API

Public

Kroki.Kroki — Module

The main Module containing the necessary types of functions for integration with a Kroki service.

Defines Base.show and corresponding Base.showable methods for different output formats and Diagram types, so they render in their most optimal form in different environments (e.g. the documentation system, Documenter output, Jupyter, etc.).

Kroki.Diagram — Type

A representation of a diagram that can be rendered by a Kroki service.

specification::AbstractString

The textual specification of the diagram

type::Symbol

The type of diagram specification (e.g. ditaa, Mermaid, PlantUML, etc.). This value is case-insensitive.

Examples

Diagram(:PlantUML, "Kroki -> Julia: Hello Julia!")

Kroki.@actdiag_str — Macro

Shorthand for instantiating actdiag Diagrams.

Kroki.@blockdiag_str — Macro

Shorthand for instantiating blockdiag Diagrams.

Kroki.@c4plantuml_str — Macro

Shorthand for instantiating c4plantuml Diagrams.

Kroki.@ditaa_str — Macro

Shorthand for instantiating ditaa Diagrams.

Kroki.@erd_str — Macro

Shorthand for instantiating erd Diagrams.

Kroki.@graphviz_str — Macro

Shorthand for instantiating graphviz Diagrams.

Kroki.@mermaid_str — Macro

Shorthand for instantiating mermaid Diagrams.

Kroki.@nomnoml_str — Macro

Shorthand for instantiating nomnoml Diagrams.

Kroki.@nwdiag_str — Macro

Shorthand for instantiating nwdiag Diagrams.

Kroki.@packetdiag_str — Macro

Shorthand for instantiating packetdiag Diagrams.

Kroki.@plantuml_str — Macro

Shorthand for instantiating plantuml Diagrams.

Kroki.@rackdiag_str — Macro

Shorthand for instantiating rackdiag Diagrams.

Kroki.@seqdiag_str — Macro

Shorthand for instantiating seqdiag Diagrams.

Kroki.@svgbob_str — Macro

Shorthand for instantiating svgbob Diagrams.

Kroki.@umlet_str — Macro

Shorthand for instantiating umlet Diagrams.

Kroki.@vega_str — Macro

Shorthand for instantiating vega Diagrams.

Kroki.@vegalite_str — Macro

Shorthand for instantiating vegalite Diagrams.

Kroki.@wavedrom_str — Macro

Shorthand for instantiating wavedrom Diagrams.

Private

Kroki.InvalidDiagramSpecificationError — Type

An Exception to be thrown when a Diagram representing an invalid specification is passed to render.

Kroki.InvalidOutputFormatError — Type

An Exception to be thrown when a Diagram is rendered to an unsupported or invalid output format.

Kroki.LIMITED_DIAGRAM_SUPPORT — Constant

Some MIME types are not supported by all diagram types, this constant contains all these limitations. The union of all values corresponds to all supported Diagram types.

Kroki.UriSafeBase64Payload — Function

UriSafeBase64Payload(diagram::Kroki.Diagram) -> String

Compresses a Diagram's specification using zlib, turning the resulting bytes into a URL-safe Base64 encoded payload (i.e. replacing + by - and / by _) to be used in communication with a Kroki service.

See the Kroki documentation for more information.

Kroki.render — Function

render(diagram::Kroki.Diagram, output_format::AbstractString) -> Array{UInt8,1}

Renders a Diagram through a Kroki service to the specified output format.

A KROKI_ENDPOINT environment variable can be set, specifying the URI of a specific instance of Kroki to use (e.g. when using a privately hosted instance). By default the publicly hosted service is used.

If the Kroki service responds with an error throws an InvalidDiagramSpecificationError or InvalidOutputFormatError if a know type of error occurs. Other errors (e.g. HTTP.ExceptionRequest.StatusError for connection errors) are propagated if they occur.